refactor: added images to extra activties and fixed spelling issue
